requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.25.2
Flask==3.1.3
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.67.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pillow==12.1.1
preshed==3.0.12
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
setuptools==82.0.1
shellingham==1.5.4
smart_open==7.5.1
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
tabulate==0.10.0
tenacity==9.1.4
thinc==8.3.10
together==2.4.0
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
types-tabulate==0.10.0.20260308
types-tqdm==4.67.3.20260303
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
websockets==16.0
Werkzeug==3.1.6
wrapt==2.1.2
